Move checkbox in to separate file.
