+++ ios/MaasAccount/Sources/MaasAccount/Requirements/Compositions/RequirementsView.swift
@@ -1,6 +1,7 @@
 import SwiftUI
 import MaasCore
 import MaasComponents
+import MaasTheme
 
 public struct RequirementsView: View, Swappable {
